--- new_task2_instructions/templates/new_task2_instructions/Results.html
@@ -1,13 +1,0 @@
-{% extends "global/Page.html" %}
-{% load otree static %}
-
-{% block title %}
-    Page title
-{% endblock %}
-
-{% block content %}
-
-    {% next_button %}
-{% endblock %}
-
-
