+++ settings.py
@@ -21,7 +21,7 @@ APP_KITS = {
             'task1_quiz_tr12',
             'task1_tr1',
             'task1_results_tr12',
-            'new_task2_instructions',
+            'task2_instructions',
             'new_task2_practice',
             'new_task2_quiz',
             'task2_tr1',
@@ -44,7 +44,7 @@ APP_KITS = {
             'task1_quiz_tr12',
             'task1_tr2',
             'task1_results_tr12',
-            'new_task2_instructions',
+            'task2_instructions',
             'new_task2_practice_chp',
             'new_task2_quiz',
             'task2_tr2',
@@ -67,7 +67,7 @@ APP_KITS = {
             'tr3_dynamic_pooled_quiz',
             'tr3_dynamic_pooled',
             'tr3_dynamic_pooled_results',
-            'new_task2_instructions',
+            'task2_instructions',
             'new_task2_practice',
             'new_task2_quiz',
             'new_task2',
@@ -100,7 +100,7 @@ SESSION_CONFIGS = [
                       'tr3_dynamic_pooled',
                       'task1_results_tr12',
                       'tr3_dynamic_pooled_results',
-                      'new_task2_instructions',
+                      'task2_instructions',
                       'new_task2_practice',
                       'new_task2_practice_chp',
                       'new_task2_quiz',
