+++ task2_instructions/templates/task2_instructions/Results.html
@@ -1,0 +1,13 @@
+{% extends "global/Page.html" %}
+{% load otree static %}
+
+{% block title %}
+    Page title
+{% endblock %}
+
+{% block content %}
+
+    {% next_button %}
+{% endblock %}
+
+
